+++ app/helpers/admin/tag_helper.rb
@@ -1,45 +1,37 @@
 # Helpers for rendering HasTagStringTags
 module Admin::TagHelper
-  def render_tags(tags, search_target: nil)
+  def render_tags(tags)
     tags.
-      map { |record_tag| render_tag(record_tag, search_target: search_target) }.
+      map { |record_tag| render_tag(record_tag) }.
       join.
       html_safe
   end
 
-  def render_tag(record_tag, search_target: nil)
+  def render_tag(record_tag)
     tag.span class: 'label label-info tag' do
-      if search_target
-        render_tag_href(record_tag, search_target)
-      else
-        render_tag_plain(record_tag)
+      if record_tag.is_a?(String)
+        record_tag = HasTagString::HasTagStringTag.from_string(record_tag)
       end
+
+      render_tag_href(record_tag)
     end
   end
 
   private
 
-  def render_tag_plain(record_tag)
-    str = record_tag.name
-    str += ":#{record_tag.value}" if record_tag.value
-    str
-  end
+  def render_tag_href(record_tag)
+    path = admin_tag_path(record_tag.name, model_type: record_tag.model_type)
 
-  def render_tag_href(record_tag, search_target)
-    str = link_to h(record_tag.name),
-                    search_target_and_query(search_target, record_tag.name)
+    str = link_to h(record_tag.name), path
 
     if record_tag.value
-      path = search_target_and_query(search_target, record_tag.name_and_value)
+      path = admin_tag_path(
+        record_tag.name_and_value, model_type: record_tag.model_type
+      )
       str += ':'
       str += link_to h(record_tag.value), path
     end
 
     str
-  end
-
-  def search_target_and_query(search_target, query)
-    query = URI::Generic.build(query: "tag=#{CGI.escape(query)}").to_s
-    search_target + query
   end
 end