+++ app/models/incoming_message.rb
@@ -37,9 +37,11 @@ require 'rexml/document'
 require 'zip'
 
 class IncomingMessage < ApplicationRecord
-  include AdminColumn
   include MessageProminence
   include CacheAttributesFromRawEmail
+  include Taggable
+
+  strip_attributes only: [:prominence_reason]
 
   MAX_ATTACHMENT_TEXT_CLIPPED = 1000000 # 1Mb ish