+++ config/routes.rb
@@ -479,6 +479,34 @@ Rails.application.routes.draw do
   resources :announcements, :only => [:destroy]
   ####
 
+  #### AdminTag controller
+  namespace :admin do
+    resources :tags, param: :tag, only: [:index, :show]
+  end
+  ####
+
+  #### AdminNote controller
+  namespace :admin do
+    resources :notes, except: [:index, :show]
+  end
+
+  direct :admin_note_parent do |note|
+    if note.notable_tag
+      admin_tag_path(tag: note.notable_tag)
+    elsif note.notable
+      url_for([:admin, note.notable])
+    else
+      admin_general_index_path
+    end
+  end
+  ####
+
+  #### Admin::PostRedirectsController
+  namespace :admin do
+    resources :post_redirects, only: [:destroy]
+  end
+  ####
+
   #### AdminPublicBody controller
   scope '/admin', :as => 'admin' do
     resources :bodies,
@@ -491,6 +519,9 @@ Rails.application.routes.draw do
         :controller => 'admin_censor_rule',
         :only => [:new, :create]
     end
+  end
+  direct :admin_public_body do |pb|
+    admin_body_path(pb)
   end
   ####
 
@@ -562,6 +593,9 @@ Rails.application.routes.draw do
         :controller => 'admin_censor_rule',
         :only => [:new, :create]
     end
+  end
+  direct :admin_info_request do |ir|
+    admin_request_path(ir)
   end
   ####